# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/transcribe:
    post:
      operationId: transcribeAudio
      tags: [voice]
      summary: Transcribe audio to text (STT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranscribeInput"
      responses:
        "200":
          description: Transcription result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeResult"
  /voice/speak:
    post:
      operationId: textToSpeech
      tags: [voice]
      summary: Convert text to speech audio (TTS) — returns SSE stream of base64 audio chunks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeakInput"
      responses:
        "200":
          description: SSE stream of base64 audio chunks
          content:
            text/event-stream: {}

  # OpenAI conversations
...
components:
  schemas:
...
    TranscribeInput:
      type: object
      properties:
        audio:
          type: string
          description: Base64-encoded audio data
      required:
        - audio
    TranscribeResult:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    SpeakInput:
      type: object
      properties:
        text:
          type: string
        voice:
          type: string
          description: Voice ID (e.g. alloy, echo, nova, shimmer)
      required:
        - text

    # OpenAI
...